LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/date-expected.txt

Inputs Date


FAIL date type support on input element assert_equals: expected "date" but got "text"
FAIL The value attribute, if specified and not empty, must have a value that is a valid date string. assert_equals: expected "2011-01-01" but got "1999-01-31"
FAIL The min attribute, if specified, must have a value that is a valid date string. assert_equals: expected "" but got "1999-1"
FAIL The max attribute, if specified, must have a value that is a valid date string. assert_equals: expected "2099-01" but got "2011-12-31"
FAIL User agents must not allow the user to set the value to a non-empty string that is not a valid date string. assert_equals: expected "" but got "invalid-date"
FAIL Number of days assert_equals: January has 31 days expected "" but got "2014-01-32"
